Refactor assembler into mapper
Basically, a RepresentationModelAssemblerSupport class is mapper. Therefore, having both an interface (Mapper) and a class (Assembler), that mostly do the same might be violating DRY principle.

From now on, the assembler will be a mapper itself. If in a future we need to extract an external mapper to separate the DTO mapping logic from the Links mapping logic, then I will add an external mapper again.