Refactored location header building
Created a private method getLocationUri that attempts to generate a location URI using a representation model link with self attribute. That way, the link is generated only once (at the assembler), and is reused for the location.

If no link with self rel is present at the representation model, then no location URI is returned by the service.