Removed AuthorProcessor
Classes implementing RepresentationModelProcessor are not meant for being autowired inside other questions assemblers. They are called automatically by Spring when a RepresentationModel is returned. 

Having that dependency was a bit overkill and made harder to test the assembler, so it is being removed and make it use a private method instead.